refactor(client): extract ./env from ./constants
