+++ client/src/contributor-spotlight/index.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { useParams } from "react-router-dom";
 import useSWR from "swr";
 
-import { CRUD_MODE } from "../constants";
+import { CRUD_MODE } from "../env";
 import { HydrationData } from "../types/hydration";
 import { GetInvolved } from "../ui/molecules/get_involved";
 import { Quote } from "../ui/molecules/quote";
