+++ client/src/document/toolbar/edit-actions.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
-import { CRUD_MODE_HOSTNAMES } from "../../constants";
+import { CRUD_MODE_HOSTNAMES } from "../../env";
 import { Source } from "../types";
 
 import "./edit-actions.scss";
