+++ client/src/homepage/featured-articles/index.tsx
@@ -1,5 +1,5 @@
 import useSWR from "swr";
-import { CRUD_MODE } from "../../constants";
+import { CRUD_MODE } from "../../env";
 import { HydrationData } from "../../types/hydration";
 
 import "./index.scss";
