+++ client/src/plus/offer-overview/offer-overview-subscribe/index.tsx
@@ -7,7 +7,7 @@ import {
   MDN_PLUS_SUBSCRIBE_5M_URL,
   MDN_PLUS_SUBSCRIBE_5Y_URL,
   MDN_PLUS_SUBSCRIBE_BASE,
-} from "../../../constants";
+} from "../../../env";
 import { SubscriptionType, UserData, useUserData } from "../../../user-context";
 import { Switch } from "../../../ui/atoms/switch";
 import { useEffect, useState } from "react";
