+++ client/src/ui/organisms/footer/index.tsx
@@ -5,7 +5,7 @@ import { useLocation } from "react-router-dom";
 
 import { ReactComponent as MDNLogo } from "../../../assets/mdn-footer-logo.svg";
 import { ReactComponent as MozLogo } from "../../../assets/moz-logo.svg";
-import { PLUS_IS_ENABLED } from "../../../constants";
+import { PLUS_IS_ENABLED } from "../../../env";
 const DARK_NAV_ROUTES = [/\/plus\/?$/i];
 
 export function Footer() {
